feat(core): add support for nunique aggregator (string only ftm)
