+++ packages/vaex-core/setup.py
@@ -108,6 +108,7 @@ extension_superutils = Extension("vaex.superutils", [
 
 extension_superagg = Extension("vaex.superagg", [
         os.path.relpath(os.path.join(dirname, "src/superagg.cpp")),
+        os.path.relpath(os.path.join(dirname, "src/agg_nunique.cpp")),
     ],
     include_dirs=[
         get_numpy_include(), get_pybind_include(),
